Shard-controller
This PR introduces prjectsveltos shard controller.

This controller watches for CAPI Cluster and SveltosCluster instances.
Internally keeps track of cluster:shard pairs.

When a cluster is part of a shard, and that is the first cluster matching
the shard, shard-controller starts projectsveltos deployments that will
manage that cluster and all clusters in the given shard.

When a cluster is not part of a shard anymore, and that was the last cluster
matching the shard, shard-controller deletes projectsveltos deployments that
used to manage clusters in that given shard.

This PR includes:
- code
- unit tests
- functional verification